Refactor : socket 씬으로 이동
- 코드리뷰
